+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Show scores setting toggle
 - Default listing type for the home tab setting
 - Import Lemmy settings: long press an account in account settings then choose the import option
+- Editing comments
 
 ## v0.4.2 - 2021-04-12